Hoist Vue translations to base-cms-leaders-program, update import
