+++ packages/leaders-program/src/utils/translations-vue.js
@@ -1,0 +1,20 @@
+export default {
+  en: {
+    visitSiteLabel: 'Visit Site',
+    viewProfileLabel: 'View Profile',
+    featuredVideosLabel: 'Featured Videos',
+    featuredProductsLabel: 'Featured Products',
+    'view more': 'View more',
+    videos: 'videos',
+    products: 'products',
+  },
+  es: {
+    visitSiteLabel: 'Visite el sitio web',
+    viewProfileLabel: 'Vea el perfil',
+    featuredVideosLabel: 'Videos destacados',
+    featuredProductsLabel: 'Productos destacados',
+    'view more': 'Vea más',
+    videos: 'videos',
+    products: 'productos',
+  },
+};
