--- packages/marko-web-leaders/translations-vue.js
@@ -1,8 +1,0 @@
-export default {
-  en: {
-    visitSiteLabel: 'Visit Site',
-  },
-  es: {
-    visitSiteLabel: 'Visite el sitio',
-  },
-};
